requirements.txt
Django==3.1.5
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
gunicorn==20.0.4
